fixed it so the tsk code is no longer in my codebase. I just reference the default installation, as well as the paths I need for the extra functions beyond the default libtsk. I also need to manually download and copy the four _i files (base_i, fs_i, img_i, vs_i) which I am using outside the default libtsk.h.
My usage is different than normal since I can recording all the properties
for each vs, fs, img, etc. which is not normal usage of calling the default
commands or functions for the commands.

This new method is cleaner and ensures any changes to the code are consistent
between releases as long as I re-download and update my tskextra files into
the respective default /usr/include/tsk/ paths.