try old method
Gonna try the old method, which was longer and more repetitive for code, but will see if it works without error.
if it does, then i will use this code and remove the cleaner method which fails. if it also fails, then i will work on fixing the cleaner method.