issue is the file has zero filesize, but it is a link to another file which isn't zero size, so it has content but my code fails.
I'll have to decide whether i ignore the link file and just hash it as zero
(i'm leaning towards), or i would have to get the info for the linked file
and then use that file size to get the content. (which i could do, but it
might be confusing)...

need to check filesize prior to starting everything else.